requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
amqp==1.4.9
anyjson==0.3.3
apipkg==1.5
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
bcrypt==3.1.7
billiard==3.3.0.23
cached-property==1.5.1
celery==3.1.26.post2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
concurrent-log-handler==0.9.9
cryptography==2.7
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0.3
django-celery==3.2.2
django-crispy-forms==1.8.1
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.10
django-stdimage==4.0.1
djangorestframework==3.10.3
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
dwebsocket==0.4.2
elasticsearch==7.9.0
et-xmlfile==1.0.1
eventlet==0.22.1
execnet==1.7.1
Flask==1.1.1
flower==0.9.2
future==0.18.2
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
har2case==0.3.1
httpie==1.0.3
httplib2==0.9.2
HttpRunner==1.5.8
idna==2.6
importlib-metadata==0.23
iniconfig==1.0.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jsonpath==0.82
jsonschema==2.6.0
kafka-python==1.4.6
kiwisolver==1.2.0
kombu==3.0.37
locustio==0.12.1
lxml==4.5.2
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.2.2
more-itertools==7.2.0
msgpack-python==0.5.6
mysqlclient==1.3.12
nb-log==3.4
numpy==1.19.0
odfpy==1.4.0
openpyxl==3.0.2
packaging==19.2
paramiko==2.6.0
pika==1.1.0
Pillow==6.2.1
pluggy==0.13.1
progressbar2==3.47.0
py==1.9.0
pyasn1==0.4.7
pycparser==2.19
pycryptodome==3.9.0
Pygments==2.4.2
pymongo==3.5.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.5
pypiwin32==223
pytest==4.5.0
pytest-assume==2.3.2
pytest-base-url==1.4.2
pytest-forked==1.3.0
pytest-html==1.19.0
pytest-metadata==1.8.0
pytest-repeat==0.7.0
pytest-rerunfailures==9.1
pytest-xdist==1.23.2
PyTestReport==0.2.1
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.2
pywin32==228
PyYAML==3.12
pyzmq==18.1.0
requests==2.22.0
requests-toolbelt==0.9.1
selenium==2.53.6
six==1.12.0
tablib==0.14.0
texttable==0.9.1
toml==0.10.1
tomorrow3==1.1.0
tornado==5.1.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
